requirements.txt
fastapi
uvicorn
python-libsbml
tellurium